fix: tighten snapshot import validation
