+++ packages/engine/src/index.ts
@@ -12,6 +12,7 @@ export { ContentManager } from './content/ContentManager.js';
 export { checkQuestionQuality } from './content/quality-filters.js';
 export { Exporter } from './export/Exporter.js';
 export { Importer } from './export/Importer.js';
+export { validateEngineSnapshot } from './export/snapshot-validation.js';
 export type { ExportBundle } from './export/Exporter.js';
 
 export type {
